Refactor: create useGuildConfig hook under /hooks/guilds folder
