--- src/contexts/Guilds/config/index.ts
@@ -1,3 +1,0 @@
-export { GuildConfigProvider } from './provider';
-export { GuildConfigContext } from './context';
-export { useGuildConfig } from './hook';
