+++ src/hooks/Guilds/useGuildConfig.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { GuildConfigContext, ConfigContext } from './context';
+import { GuildConfigContext, ConfigContext } from '../../contexts/Guilds';
 
 export const useGuildConfig = () =>
   React.useContext<ConfigContext>(GuildConfigContext);
